Convert to use Reader
Instead of passing in L10n etc. everywhere, have this information in a
state value that we can query via the Reader monad's machinery. This
hugely simplifies type signatures and makes the code more easily
extensible. For example, we can add a localization map for this
program's messages, enabling us to internationalize it.