Refactor to use a proper message system
Instead of using Doc and hand-hacking messages, we use the Shakespeare
library to emit messages via a message system. A message is a
constructor of a message type, which gets converted to Text via a
function that's aware of what language is wanted.

This makes some messages harder: new leaders and rulers for example have
many optional clauses and accounting for them all requires a rather
complex message.