Use quasiquoter
EU4.Common now uses the quasiquoter as extensively as possible. This
version's output has been verified by recursive diff to be the same as
that of 0.3.4.1.

The quasiquoter was renamed again, because the new name turned out to
also be a common variable name.

This commit additionally broke some of the longer type signatures over
several lines.