Refactor to prepare for further code reuse
Game settings are wrapped in an existential type and the hard work is
punted to class instances. Each game has a class that implements stuff
specific to that game. It should be possible to move most functions that
don't use these classes to a common module usable by other games
(towards #2).