Improve internal documentation
Since I'm rather confused about the structure of the program at this
point, I decided to leverage haddock to document it. This is step 1:
providing haddocks for the toplevel modules.

Stack currently doesn't build haddocks for executables, so a second step
will be to restructure the program as a library, which actually isn't
too difficult.