Make pronouns context sensitive, + more messages
"ROOT" was previously translated as "(Our country)", which is not
correct in cases where ROOT refers to a province, trade node etc. This
patch makes interpretation of pronouns dependent on the context in which
they occur.

FROM is, unfortunately, inscrutable as it needs additional context to
interpret, so that is still left untranslated. PREV is similarly also
inscrutible where it refers to an object brought into scope by
event_target or the like.

Various handlers now take an "expected context" parameter, which is the
type of the statement's RHS. This is important in determining the
meaning of pronouns such as ROOT and PREV, which in many cases refer to
objects of different type (for example, checking whether something's
controller is a province means that province's owner).