Partial update for Dharma
As well as a few content additions for 1.27, this commit updates the
parser to permit arbitrarily many tags in an atom sequence, since these
are needed for Imperator: Rome. E.g. that game has such entities on the
LHS as hidden:scope:target.employer = { ... }, which we need to be able
to handle.

Version number for EU4 has been bumped to 1.28, though it hasn't been
tested with that version.